Refactor code generator: update include paths and remove executeMain method
